Use test helpers
