+++ src/libraries/System.Memory.Data/tests/BinaryDataTests.cs
@@ -431,73 +431,6 @@ namespace System.Tests
             var data = BinaryData.FromStream(new OverFlowStream(offset: int.MaxValue - 1000));
         }
 
-        [Fact]
-        public async Task CanCreateBinaryDataFromFile()
-        {
-            byte[] buffer = "some data"u8.ToArray();
-            string path = Path.GetTempFileName();
-            try
-            {
-                File.WriteAllBytes(path, buffer);
-                BinaryData data = BinaryData.FromFile(path);
-                Assert.Equal(buffer, data.ToArray());
-
-                byte[] output = new byte[buffer.Length];
-                var outputStream = data.ToStream();
-                outputStream.Read(output, 0, (int)outputStream.Length);
-                Assert.Equal(buffer, output);
-
-                data = await BinaryData.FromFileAsync(path);
-                Assert.Equal(buffer, data.ToArray());
-
-                outputStream = data.ToStream();
-                outputStream.Read(output, 0, (int)outputStream.Length);
-                Assert.Equal(buffer, output);
-            }
-            finally
-            {
-                File.Delete(path);
-            }
-        }
-
-        [Theory]
-        [InlineData(null)]
-        [InlineData("")]
-        [InlineData(MediaTypeNames.Application.Soap)]
-        public async Task CanCreateBinaryDataFromFileWithMediaType(string? mediaType)
-        {
-            byte[] buffer = "some data"u8.ToArray();
-            string path = Path.GetTempFileName();
-            try
-            {
-                File.WriteAllBytes(path, buffer);
-                BinaryData data = BinaryData.FromFile(path, mediaType);
-                Assert.Equal(buffer, data.ToArray());
-                Assert.Equal(mediaType, data.MediaType);
-
-                byte[] output = new byte[buffer.Length];
-                var outputStream = data.ToStream();
-                outputStream.Read(output, 0, (int)outputStream.Length);
-                Assert.Equal(buffer, output);
-
-                data = await BinaryData.FromFileAsync(path, mediaType);
-                Assert.Equal(buffer, data.ToArray());
-                Assert.Equal(mediaType, data.MediaType);
-
-                outputStream = data.ToStream();
-                outputStream.Read(output, 0, (int)outputStream.Length);
-                Assert.Equal(buffer, output);
-
-                //changing the backing buffer should not affect the BD instance
-                buffer[3] = (byte)'z';
-                Assert.NotEqual(buffer, data.ToMemory().ToArray());
-            }
-            finally
-            {
-                File.Delete(path);
-            }
-        }
-
         [ConditionalFact(typeof(PlatformDetection), nameof(PlatformDetection.IsNotBuiltWithAggressiveTrimming))]
         public void CanCreateBinaryDataFromCustomType()
         {
@@ -564,67 +497,30 @@ namespace System.Tests
         [Fact]
         public async Task CreateThrowsOnNullStream()
         {
-            var ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromStream(null));
-            Assert.Contains("stream", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromStream(null, null));
-            Assert.Contains("stream", ex.Message);
-
-            ex = await Assert.ThrowsAsync<ArgumentNullException>(() => BinaryData.FromStreamAsync(null));
-            Assert.Contains("stream", ex.Message);
-
-            ex = await Assert.ThrowsAsync<ArgumentNullException>(() => BinaryData.FromStreamAsync(null, null));
-            Assert.Contains("stream", ex.Message);
-        }
-
-        [Fact]
-        public async Task CreateFromFileThrowsOnNullPath()
-        {
-            var ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromFile(null));
-            Assert.Contains("path", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromFile(null, null));
-            Assert.Contains("path", ex.Message);
-
-            ex = await Assert.ThrowsAsync<ArgumentNullException>(() => BinaryData.FromFileAsync(null));
-            Assert.Contains("path", ex.Message);
-
-            ex = await Assert.ThrowsAsync<ArgumentNullException>(() => BinaryData.FromFileAsync(null, null));
-            Assert.Contains("path", ex.Message);
+            AssertExtensions.Throws<ArgumentNullException>("stream", () => BinaryData.FromStream(null));
+            AssertExtensions.Throws<ArgumentNullException>("stream", () => BinaryData.FromStream(null, null));
+            await AssertExtensions.ThrowsAsync<ArgumentNullException>("stream", () => BinaryData.FromStreamAsync(null));
+            await AssertExtensions.ThrowsAsync<ArgumentNullException>("stream", () => BinaryData.FromStreamAsync(null, null));
         }
 
         [Fact]
         public void CreateThrowsOnNullString()
         {
             string payload = null;
-            var ex = Assert.Throws<ArgumentNullException>(() => new BinaryData(payload));
-            Assert.Contains("data", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => new BinaryData(payload, null));
-            Assert.Contains("data", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromString(payload));
-            Assert.Contains("data", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromString(payload, null));
-            Assert.Contains("data", ex.Message);
+            var ex = AssertExtensions.Throws<ArgumentNullException>("data", () => new BinaryData(payload));
+            AssertExtensions.Throws<ArgumentNullException>("data", () => new BinaryData(payload, null));
+            AssertExtensions.Throws<ArgumentNullException>("data", () => BinaryData.FromString(payload));
+            AssertExtensions.Throws<ArgumentNullException>("data", () => BinaryData.FromString(payload, null));
         }
 
         [Fact]
         public void CreateThrowsOnNullArray()
         {
             byte[] payload = null;
-            var ex = Assert.Throws<ArgumentNullException>(() => new BinaryData(payload));
-            Assert.Contains("data", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => new BinaryData(payload, null));
-            Assert.Contains("data", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromBytes(null));
-            Assert.Contains("data", ex.Message);
-
-            ex = Assert.Throws<ArgumentNullException>(() => BinaryData.FromBytes(null, null));
-            Assert.Contains("data", ex.Message);
+            var ex = AssertExtensions.Throws<ArgumentNullException>("data", () => new BinaryData(payload));
+            AssertExtensions.Throws<ArgumentNullException>("data", () => new BinaryData(payload, null));
+            AssertExtensions.Throws<ArgumentNullException>("data", () => BinaryData.FromBytes(payload));
+            AssertExtensions.Throws<ArgumentNullException>("data", () => BinaryData.FromBytes(payload, null));
         }
 
         [ConditionalFact(typeof(PlatformDetection), nameof(PlatformDetection.IsNotBuiltWithAggressiveTrimming))]
